centralize decision metrics
